+++ functions.ts
@@ -3,3 +3,4 @@ export { validateRows } from "./src/validators/row-validator/validateRows";
 export { validateColumns } from "./src/validators/column-validator/validateColumns";
 export { validateThermos } from "./src/validators/thermo-validator/validateThermos";
 export { validateKropki } from "./src/validators/kropki-validator/validateKropki";
+export { validateDiagonals } from "./src/validators/diagonal-validator/validateDiagonal";
