+++ src/validators/diagonal-validator/validateDiagonal.ts
